Code cleanup -Move pg listener out for main app.ts
